requirement.txt
alembic==0.9.9
Babel==2.6.0
backcall==0.1.0
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
DBUtils==1.2
decorator==4.3.0
defusedxml==0.5.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.0
Flask-OpenID==1.2.5
flask-paginate==0.5.1
Flask-Script==2.0.6
flask-shell==0.1.3
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysql-connector==2.1.6
parso==0.3.1
pbr==4.0.4
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==1.0.15
Pygments==2.2.0
PyMySQL==0.8.1
python-dateutil==2.7.3
python-editor==1.0.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.4
qiniu==7.2.2
requests==2.19.1
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
